feat: add new TypeScript interfaces for charts, exports, images, and locations
