+++ src/app/admin/msme/[sectorName]/page.tsx
@@ -21,7 +21,7 @@ import {
 import { useMSMEContext } from "@/contexts/MSMEContext";
 import AdminAddMSMEModal from "@/components/modals/AdminAddMSMEModal";
 import AdminEditMSMEModal from "@/components/modals/AdminEditMSMEModal";
-import type { MSME } from "@/types/superadmin";
+import type { MSME } from "@/types/MSME";
 import { MSMETableView } from "@/components/admin/MSMETable";
 import { cn } from "@/lib/utils";
 import {
