+++ src/app/msme/[id]/page.tsx
@@ -17,10 +17,7 @@ import type { CarouselApi } from "@/components/ui/carousel";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { Spinner } from "@/components/ui/spinner";
-
-interface MSMEWithSectorName extends MSME {
-  sectorName: string | null;
-}
+import type { MSMEWithSectorName } from "@/types/MSME";
 
 export default function MSMEPage({ params }: { params: { id: string } }) {
   const [MSME, setMSME] = useState<MSMEWithSectorName>();
