+++ src/components/modals/AddMSMEModal.tsx
@@ -28,10 +28,7 @@ import { uploadImage } from "@/utils/supabase/storage";
 import { toast } from "sonner";
 import ImageCropModal from "@/components/modals/ImageCropModal";
 import { GoogleMap, Marker, useJsApiLoader } from "@react-google-maps/api";
-interface AddMSMEModalProps {
-  isOpen: boolean;
-  onClose: () => void;
-}
+import type { AddMSMEModalProps } from "@/types/MSME";
 
 export default function AddMSMEModal({ isOpen, onClose }: AddMSMEModalProps) {
   const { sectors, handleAddMSME } = useMSMEContext();
