+++ src/types/APIResponse.ts
@@ -1,0 +1,6 @@
+import { SignupSector } from "@/types/sector";
+
+export interface ApiResponse {
+  sectors: SignupSector[];
+  error?: string;
+}
