+++ src/types/charts.ts
@@ -1,0 +1,5 @@
+export interface BarChart2Props {
+  data: Array<{ name: string; uv: number }>;
+  label: string;
+  color?: string;
+}
