+++ src/types/export.ts
@@ -1,0 +1,4 @@
+export interface ExportsLineChartProps {
+  data: { month: string; exports: number }[];
+  totalExports: number;
+}
