+++ src/types/image.ts
@@ -1,0 +1,5 @@
+export interface ImageCropModalProps {
+  isOpen: boolean;
+  onClose: () => void;
+  onCropComplete: (file: File) => void;
+}
